requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
gunicorn==23.0.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
cryptography==43.0.1
weasyprint==62.3
redis==5.1.1
structlog==24.4.0
opentelemetry-sdk==1.26.0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-instrumentation-sqlalchemy==0.47b0
openai==1.54.1
tiktoken==0.8.0
tenacity==9.0.0

